Move the helper methods to a shared utility class
